Fix warnings about comparisons of signed and unsigned integers

See merge request amd/rocPRIM!38

+++ test/test_rocprim_hc_warp_sort.cpp
@@ -96,7 +96,7 @@ TYPED_TEST(RocprimWarpSortShuffleBasedTests, SortInt)
     );
 
     d_output.synchronize();
-    for(int i = 0; i < output.size(); i++)
+    for(size_t i = 0; i < output.size(); i++)
     {
         EXPECT_EQ(output[i], expected[i]);
     }
@@ -147,7 +147,7 @@ TYPED_TEST(RocprimWarpSortShuffleBasedTests, SortKeyInt)
 
     d_output_key.synchronize();
     d_output_value.synchronize();
-    for(int i = 0; i < expected.size(); i++)
+    for(size_t i = 0; i < expected.size(); i++)
     {
         EXPECT_EQ(d_output_key[i], expected[i].first);
         EXPECT_EQ(d_output_value[i], expected[i].second);
